# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Blockchain Wallet Tracing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface WalletInfo {
  address: string;
  chain: string;
  balance: string;
  balanceUsd: number;
  transactionCount: number;
  firstSeen: string | null;
  lastSeen: string | null;
  tags: string[];
  riskScore: number | null;
  isContract: boolean;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  success: "success",
  failed: "failed",
  pending: "pending",
} as const;

export type TransactionDirection =
  (typeof TransactionDirection)[keyof typeof TransactionDirection];

export const TransactionDirection = {
  in: "in",
  out: "out",
  self: "self",
} as const;

export interface Transaction {
  hash: string;
  from: string;
  to: string | null;
  value: string;
  valueUsd: number;
  fee: string;
  feeUsd: number;
  timestamp: string;
  blockNumber: number;
  status: TransactionStatus;
  direction: TransactionDirection;
  tokenSymbol: string | null;
  tokenName: string | null;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface WalletNode {
  address: string;
  label: string | null;
  balance: string;
  transactionCount: number;
  isContract: boolean;
  riskScore: number | null;
}

export interface WalletEdge {
  from: string;
  to: string;
  totalValue: string;
  totalValueUsd: number;
  transactionCount: number;
  lastSeen: string;
}

export interface WalletConnections {
  nodes: WalletNode[];
  edges: WalletEdge[];
  centerAddress: string;
}

export interface RecentSearch {
  id: number;
  address: string;
  chain: string;
  searchedAt: string;
  label: string | null;
}

export interface SaveSearchBody {
  address: string;
  chain: string;
  label?: string | null;
}

export type SearchStatsPopularWalletsItem = {
  address: string;
  chain: string;
  searchCount: number;
  label: string | null;
};

export type SearchStatsRecentActivityItem = {
  date: string;
  count: number;
};

export interface SearchStats {
  totalSearches: number;
  uniqueWallets: number;
  popularWallets: SearchStatsPopularWalletsItem[];
  recentActivity: SearchStatsRecentActivityItem[];
}

export type GetWalletParams = {
  chain?: GetWalletChain;
};

export type GetWalletChain =
  (typeof GetWalletChain)[keyof typeof GetWalletChain];

export const GetWalletChain = {
  ethereum: "ethereum",
  bitcoin: "bitcoin",
  polygon: "polygon",
  bsc: "bsc",
} as const;

export type GetWalletTransactionsParams = {
  chain?: GetWalletTransactionsChain;
  page?: number;
  limit?: number;
};

export type GetWalletTransactionsChain =
  (typeof GetWalletTransactionsChain)[keyof typeof GetWalletTransactionsChain];

export const GetWalletTransactionsChain = {
  ethereum: "ethereum",
  bitcoin: "bitcoin",
  polygon: "polygon",
  bsc: "bsc",
} as const;

export type GetWalletConnectionsParams = {
  chain?: GetWalletConnectionsChain;
  /**
   * @minimum 1
   * @maximum 3
   */
  depth?: number;
};

export type GetWalletConnectionsChain =
  (typeof GetWalletConnectionsChain)[keyof typeof GetWalletConnectionsChain];

export const GetWalletConnectionsChain = {
  ethereum: "ethereum",
  bitcoin: "bitcoin",
  polygon: "polygon",
  bsc: "bsc",
} as const;