# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Blockchain Wallet Tracing API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get wallet info and balance
 */
export const GetWalletParams = zod.object({
  address: zod.coerce.string(),
});

export const getWalletQueryChainDefault = `ethereum`;

export const GetWalletQueryParams = zod.object({
  chain: zod
    .enum(["ethereum", "bitcoin", "polygon", "bsc"])
    .default(getWalletQueryChainDefault),
});

export const GetWalletResponse = zod.object({
  address: zod.string(),
  chain: zod.string(),
  balance: zod.string(),
  balanceUsd: zod.number(),
  transactionCount: zod.number(),
  firstSeen: zod.string().nullable(),
  lastSeen: zod.string().nullable(),
  tags: zod.array(zod.string()),
  riskScore: zod.number().nullable(),
  isContract: zod.boolean(),
});

/**
 * @summary Get wallet transaction history
 */
export const GetWalletTransactionsParams = zod.object({
  address: zod.coerce.string(),
});

export const getWalletTransactionsQueryChainDefault = `ethereum`;
export const getWalletTransactionsQueryPageDefault = 1;
export const getWalletTransactionsQueryLimitDefault = 20;

export const GetWalletTransactionsQueryParams = zod.object({
  chain: zod
    .enum(["ethereum", "bitcoin", "polygon", "bsc"])
    .default(getWalletTransactionsQueryChainDefault),
  page: zod.coerce.number().default(getWalletTransactionsQueryPageDefault),
  limit: zod.coerce.number().default(getWalletTransactionsQueryLimitDefault),
});

export const GetWalletTransactionsResponse = zod.object({
  transactions: zod.array(
    zod.object({
      hash: zod.string(),
      from: zod.string(),
      to: zod.string().nullable(),
      value: zod.string(),
      valueUsd: zod.number(),
      fee: zod.string(),
      feeUsd: zod.number(),
      timestamp: zod.string(),
      blockNumber: zod.number(),
      status: zod.enum(["success", "failed", "pending"]),
      direction: zod.enum(["in", "out", "self"]),
      tokenSymbol: zod.string().nullable(),
      tokenName: zod.string().nullable(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Get connected wallets (traced relationships)
 */
export const GetWalletConnectionsParams = zod.object({
  address: zod.coerce.string(),
});

export const getWalletConnectionsQueryChainDefault = `ethereum`;
export const getWalletConnectionsQueryDepthDefault = 1;
export const getWalletConnectionsQueryDepthMax = 3;

export const GetWalletConnectionsQueryParams = zod.object({
  chain: zod
    .enum(["ethereum", "bitcoin", "polygon", "bsc"])
    .default(getWalletConnectionsQueryChainDefault),
  depth: zod.coerce
    .number()
    .min(1)
    .max(getWalletConnectionsQueryDepthMax)
    .default(getWalletConnectionsQueryDepthDefault),
});

export const GetWalletConnectionsResponse = zod.object({
  nodes: zod.array(
    zod.object({
      address: zod.string(),
      label: zod.string().nullable(),
      balance: zod.string(),
      transactionCount: zod.number(),
      isContract: zod.boolean(),
      riskScore: zod.number().nullable(),
    }),
  ),
  edges: zod.array(
    zod.object({
      from: zod.string(),
      to: zod.string(),
      totalValue: zod.string(),
      totalValueUsd: zod.number(),
      transactionCount: zod.number(),
      lastSeen: zod.string(),
    }),
  ),
  centerAddress: zod.string(),
});

/**
 * @summary Get recent wallet searches
 */
export const GetRecentSearchesResponseItem = zod.object({
  id: zod.number(),
  address: zod.string(),
  chain: zod.string(),
  searchedAt: zod.string(),
  label: zod.string().nullable(),
});
export const GetRecentSearchesResponse = zod.array(
  GetRecentSearchesResponseItem,
);

/**
 * @summary Save a wallet search
 */
export const SaveSearchBody = zod.object({
  address: zod.string(),
  chain: zod.string(),
  label: zod.string().nullish(),
});

/**
 * @summary Get search statistics and popular wallets
 */
export const GetSearchStatsResponse = zod.object({
  totalSearches: zod.number(),
  uniqueWallets: zod.number(),
  popularWallets: zod.array(
    zod.object({
      address: zod.string(),
      chain: zod.string(),
      searchCount: zod.number(),
      label: zod.string().nullable(),
    }),
  ),
  recentActivity: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
    }),
  ),
});